LayoutTests/css3/color-filters/color-filter-exposed-if-enabled-expected.txt

Tests that -apple-color-filter is exposed when the feature is enabled

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS '-apple-color-filter' in document.documentElement.style is true
PASS document.documentElement.style['-apple-color-filter'] is not undefined
PASS document.documentElement.style.getPropertyValue('-apple-color-filter') is 'contrast(1)'
PASS document.documentElement.style.getPropertyPriority('-apple-color-filter') is 'important'
PASS document.documentElement.style.getPropertyCSSValue('-apple-color-filter').length is 1
PASS document.documentElement.style.getPropertyCSSValue('-apple-color-filter')[0].cssValueType is 3
PASS document.documentElement.style.getPropertyCSSValue('-apple-color-filter')[0].cssText is 'contrast(1)'
PASS document.documentElement.style.removeProperty('-apple-color-filter') is 'contrast(1)'
PASS document.documentElement.style.getPropertyValue('-apple-color-filter') is ""
PASS document.documentElement.style.setProperty('-apple-color-filter', 'contrast(1)', '') is undefined.
PASS document.documentElement.style.getPropertyValue('-apple-color-filter') is 'contrast(1)'
PASS '-apple-color-filter' in getComputedStyle(document.documentElement) is true
PASS getComputedStyle(document.documentElement)['-apple-color-filter'] is not undefined
PASS getComputedStyle(document.documentElement).getPropertyValue('-apple-color-filter') is 'contrast(1)'
PASS getComputedStyle(document.documentElement).getPropertyPriority('-apple-color-filter') is ""
PASS getComputedStyle(document.documentElement).getPropertyCSSValue('-apple-color-filter').length is 1
PASS getComputedStyle(document.documentElement).getPropertyCSSValue('-apple-color-filter')[0].cssValueType is 3
PASS getComputedStyle(document.documentElement).getPropertyCSSValue('-apple-color-filter')[0].cssText is 'contrast(1)'
PASS getComputedStyle(document.documentElement).removeProperty('-apple-color-filter') threw exception NoModificationAllowedError: The object can not be modified..
PASS getComputedStyle(document.documentElement).getPropertyValue('-apple-color-filter') is 'contrast(1)'
PASS getComputedStyle(document.documentElement).setProperty('-apple-color-filter', 'contrast(1)', '') threw exception NoModificationAllowedError: The object can not be modified..
PASS getComputedStyle(document.documentElement).getPropertyValue('-apple-color-filter') is 'contrast(1)'
PASS 'AppleColorFilter' in document.documentElement.style is true
PASS document.documentElement.style['AppleColorFilter'] is not undefined
PASS 'AppleColorFilter' in getComputedStyle(document.documentElement) is true
PASS getComputedStyle(document.documentElement)['AppleColorFilter'] is not undefined
PASS CSS.supports('-apple-color-filter: contrast(1)') is true
PASS CSS.supports('-apple-color-filter: inherit') is true
PASS CSS.supports('-apple-color-filter', 'contrast(1)') is true
PASS CSS.supports('-apple-color-filter', 'inherit') is true
PASS getComputedStyle(document.getElementById('test1')).color is 'rgb(255, 165, 0)'
PASS getComputedStyle(document.getElementById('test2')).color is 'rgb(255, 165, 0)'
PASS successfullyParsed is true

TEST COMPLETE
